fix update job without tags

+++ backend/storage/jobs.js
@@ -163,18 +163,22 @@ var update = module.exports.update = Promise.method(function (context, id, attri
                     });
             })
             .then(function() {
-                return models.JobTag.destroy({where:{ jobId: locals.job.dataValues.id}, transaction: tx});
+                if (locals.attrs.tags) {
+                    return models.JobTag.destroy({where: {jobId: locals.job.dataValues.id}, transaction: tx});
+                }
             })
             .then(function(){
-                var tags;
-                var arrayData = locals.attrs.tags;
-                arrayData.forEach(function(tag) {
-                    tags = {
-                        jobId: locals.job.dataValues.id,
-                        tag: tag.toString()
-                    }
-                    models.JobTag.create(tags);
-                });
+                if (locals.attrs.tags) {
+                    var tags;
+                    var arrayData = locals.attrs.tags;
+                    arrayData.forEach(function (tag) {
+                        tags = {
+                            jobId: locals.job.dataValues.id,
+                            tag: tag.toString()
+                        }
+                        models.JobTag.create(tags);
+                    });
+                }
             });
         })
         .catch(function (err) {
